requirements-test.txt
bitshares
pyyaml
pytest
coverage
